added total price, ordertype(still working on it)

+++ src/main/java/edu/metrostate/Item.java
@@ -19,7 +19,7 @@ public class Item {
 
     //setters
     public void setName(String name) {this.name = name;}
-    public void setPrice(float price) {this.price = price;}
+    public void setPrice(double price) {this.price = price;}
     public void setQuantity(int quantity) {this.quantity = quantity;}
 
 }
